kyoutei_ai/requirements.txt
# ボートレース予測AI システム依存パッケージ

# データ処理
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0
pyarrow>=14.0.0
polars>=0.19.0

# データ収集・スクレイピング
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
playwright>=1.40.0
fake-useragent>=1.4.0
cloudscraper>=1.2.71
httpx>=0.25.0
curl-cffi>=0.5.10

# LZHアーカイブ処理
unlzh>=0.1.1
py7zr>=0.20.0

# 機械学習
lightgbm>=4.1.0
xgboost>=2.0.0
scikit-learn>=1.3.0
optuna>=3.4.0
catboost>=1.2.0

# 可視化
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
streamlit>=1.28.0

# LLM連携
ollama>=0.1.0
openai>=1.3.0
langchain>=0.0.340
discord.py>=2.3.0

# 数値最適化
scipy>=1.11.0
cvxpy>=1.4.0

# ユーティリティ
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
pydantic>=2.5.0
tenacity>=8.2.0
schedule>=1.2.0
typer>=0.9.0
rich>=13.7.0

# 日本語処理
mojimoji>=0.0.12
jaconv>=0.3.0

# テスト
pytest>=7.4.0
pytest-asyncio>=0.21.0
